refactor/fix comments to PR
